requirements.txt
asgiref==3.7.2
cloudinary==1.36.0
crispy-bootstrap5==0.7
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
django==4.2.7
django-allauth==0.57.0
django-crispy-forms==2.1
django-summernote==0.8.20.0
gunicorn==20.1.0
oauthlib==3.2.2
Pillow==8.3.2
psycopg2==2.9.6
PyJWT==2.8.0
python3-openid==3.2.0
python-decouple==3.4
python-dotenv==0.19.1
requests-oauthlib==1.3.1
stripe==8.8.0.
sqlparse==0.4.4
urllib3==1.26.18
whitenoise==5.3.0
django[argon2]